Move setupUsers to test-helpers
Will facilitate creating new tests like this one for the auth
middleware.
